move application code under src/ folder
