+++ tests/constants_test.py
@@ -1,6 +1,6 @@
 import unittest
 
-from constants import (DAY, MONTH, MONTHS_WITH_31_DAYS, PIVOT_DATE, PIVOT_DAY,
+from src.constants import (DAY, MONTH, MONTHS_WITH_31_DAYS, PIVOT_DATE, PIVOT_DAY,
                        Date)
