+++ tests/utils_test.py
@@ -1,8 +1,8 @@
 import unittest
 
-from constants import MONTH, PIVOT_DATE, Date
-from exceptions import InvalidDateFormat
-from utils import (countLeapYears, dateValidator, getActualDaysInMonth,
+from src.constants import MONTH, PIVOT_DATE, Date
+from src.exceptions import InvalidDateFormat
+from src.utils import (countLeapYears, dateValidator, getActualDaysInMonth,
                    getDefaultDaysInMonth, isLeapYear, numDaysBetweenDates)
